notifier: move BuildManagedInhibitionRules to merge package
